shuffle tests around
